backend.old/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
passlib[bcrypt]==1.7.4
python-multipart
pydantic
pydantic[email]
python-dotenv==1.0.0
python-dateutil
aiogram
